Add example of GIF processing using ImageMagick
This commit adds an example of processing gifs using ImageMagick and gifsicle.
The workflow creates a PVC that is used for locally storing the downloaded
images and then running 'convert' and 'gifsicle' on these images.

The output gets written in the PV itself in a "output" directory. This directory
is then considered as an artifact and uploaded to the cloud.