Allow containers to be retried. (#661)
* Allow containers to be retried.

This commit allows `container` templates to be retried. The user
simply adds a 'retries' section to the templated. Currently, this
section only has a 'limit' field that has the number of times the
container should be retried in case of failure.

Subsequent commits will add more features such as retries policies,
retries on specific errors, erc.

When a container template has a retries section, the workflow
controller adds an intermediate node to the workflow. This node
acts as a parent of all the retries.

The argo command line tool's 'get' command correctly shows the
intermediate node and the child nodes along with their pod
names and status'.

Added unit tests that check the various state transitions of the
intermediate node based on the status of the child nodes (which
actually execute the pod).

Testing Done.

* Unit tests succeeded.

* Ran simple workflow that had retries. It completed and status
  of the nodes was correct.

* Ran workflows with parallel steps, each of which had containers
  with retries. The workflow completed correctly and the retries
  were applied to each individual container correctly.

* Ran other workflows which did not have retries. The completed
  correctly.

* `argo get <wf-name>` shows the retries and the pod names.

* Incorporate review comments.

* Rename example files for retries.

* Update autogenerated code to include RetryStrategy.

* Call processNodeRetries from executeTemplate.

Earlier, this was getting called during podReconciliation. Calling
this from executeTemplate is better since other non-leaf nodes also
getting processed in executeTemplate. podReconciliation should
only process leaf-nodes (which are based on execution of pods).

* Reduce failure probabilities of retry examples.

This will prevent spurious failures during e2e tests. Users wanting
to experiment with retries will have to explicitly change the
failure probability in the yamls.