Switch to Updating workflow resources instead of Patching (resolves #686)
Relying on a JSON merge patch to persist workflow updates has been shown to be unreliable.
This change switches us to use the Update mechanism, which may not be as performant, and
will on occasion cause the controller to redo work (upon resource conflicts), but it will
likely be more reliable way of persisting workflow resource changes.