Trim spaces from aws keys
when using aws keys, an extra newline was being added to both
secret/access key. I could not figure out what was adding this newline,
and feel confident that I was not adding it during my base64 encoding
process, so I added a patch to remove newlines so that I could evaluate
argo with S3.