Omit empty status fields.
Primary motivation is to avoid an empty string phase value, which is
awkward to map to JSON in languages other than Go, and surprising as
it only arises in ADDED events.

This is could be a disruptive change for workflow clients expecting the
empty value, but not sure this is likely in practice.

A newly added workflow now gives:
```json
    "status": {
        "finishedAt": null,
        "startedAt": null
    }
```

The time values already have omitempty but as they're structs this
doesn't help. The nulls here are consistent with time serialization
elsewhere in Kubernetes so probably best left as is.