Minimal shell completion support (#807)
* Minimal shell completion support.

The subcommand follows the pattern of kubectl's completion support,
though that has considerable extensions.

zsh support in cobra is minimal at the moment and doesn't support flag
completion. However, there are open PRs against it, so it doesn't seem
worth going the Kubernetes route which modifies the bash output to make
it suitable for zsh.

An obvious next step would be completing to workflow names. I doubt this
would be as useful as pod completions in kubectl, as there isn't often
a useful prefix (if you run many of the same type of workflow).

* Tweak language.

* Hopefully fix lint error.

I've restructured the flow so there's only one point to check the error.