edit Argo license info so that GitHub recognizes it
GitHub uses a library called Licensee to identify a project's license
type. It shows this information in the status bar and via the API if it
can unambiguously identify the license.

This commit removes the COPYRIGHT file and updates the LICENSE file to
reference "The Argo Authors". This all allows Licensee to successfully
identify the license type of Argo's codebase as Apache 2.0.

Signed-off-by: Andrea Kao <eirinikos@gmail.com>