Add parameter-file support
For workflows that have many parameters, or cases where the workflows should be run with slightly different requirements, it can be useful to have a file containing parameters instead of submitting via the command line.

This commit adds support for the `--parameter-file` (`-f`) command line option to provide a new-line-delimited file of parameters.  These parameters can still be overridden by the `--parameter` (`-p`) flags.

Comments in parameter files are also supported and denoted by a `#` character.

See #796 for more information