fix(docker): fix streaming of combined stdout/stderr
This PR fixes #2261. Multireader concactinates, not combines, the two
streams, this meant that we blocked on stdin if stderr was not
completely buffered.

In addition, this implementation ensures that we call Wait() on the
command, and any error is return by Close(). Godoc for
exec.Cmd.Std(out|err)Pipe, confirms that we do not need to close those
pipes.

This PR also ensures that we do not leak go routines in the even that
cmd.Start() fails.