feat(UI): Move Workflow parameters to top of submit
When uploading a workflow, or submitting a template, it is more likely that one just wants to submit the workflow by simply modifying parameter values. This also makes Workflow / WorkflowTemplate submissisons less scary for users not familiar with the Worklow YAML itself, as they can simply focus on the top part of the screen and ignore the YAML edit box below.