feat(UI): Submit WFTs with entrypoint as Workflow
WorkflowTemplates that define an entrypoint can be treated as a regular Workflow. For classic WorkflowTemplates without entrypoint, a generator generates a Workflow spec that references the WorkflowTemplate in the WorkflowTemplate submit UI. This is however not necessary if the WorkflowTemplate defines an entrypoint, and by loading the WorkflowTemplate into the submission form the user can get access to the argument input fields.