feat(manifests): add name on workflow-controller-metrics service port
While adding a name is optional from a Kubernetes standpoint since there
is only one port listed, the named port makes using Prometheus easier.
There's a targetPort option for Prometheus' ServiceMonitor Custom
Resource, but this targetPort doesn't relate to the service's
targetPort, but in fact the containers' targetPort pointed at by the
service [1]. And since the workflow-controller deployment doesn't
actually explicitly list the ports the targetPort configuration on a
ServiceMonitor does nothing. So having the named port on the service's
port makes it much easier to use for Prometheus.

1 - https://github.com/coreos/prometheus-operator/issues/2515