test(recurse mode): add e2e test for new recurseMode API
Note: the "recurseMode" key is only used if it is set at the template input level rather than task
argument. This behaviour is confirmed by the "artifact-mode-recurse-argument" step which shows that
despite setting "recurseMode" to "true" in the task arguments the mode of the contents of the
"message" artifact folder are not changed to "777".