fix(ui): handle logging disconnects gracefully
When the API finishes streaming logs from a given container, it's normal for
the server to close the connection which results in an error being emitted in the browser.
When one of those errors is received, we check whether or not the node is still pending or
running which allows us to differentiate between a disconnection that happens due to a network
error, and a disconnect that happens due to the completion of log streaming.

Fixes #4117