feat: Support per-output parameter aggregation
Currently Argo only supports the `{{tasks.<TASKNAME>.outputs.parameters.<NAME>}}` placeholder for the outputs of normal nodes. (For the outputs of the loop nodes there is the `{{tasks.<TASKNAME>.outputs.parameters}}` placeholder, which is replaced with JSON-serialized list of parameter value maps.

This PR adds support for the `{{tasks.<TASKNAME>.outputs.parameters.<NAME>}}` placeholder for the outputs of the loop nodes. The placeholder is replaced with JSON-serialized list of parameter values.

Signed-off-by: Alexey Volkov <alexey.volkov@ark-kun.com>