fix(server): use the correct name when downloading artifacts
When artifacts are uploaded to whatever repository is in use, they're stored
under a key that includes both the intended file name as well as the extension
to use based on whether or not compression was used when it was created.

Use that key as the file name when returning artifacts for download rather than
just "filename.tgz", that way we don't incorrectly identify non-compressed artifacts
as compressed tarballs.

Fixes #4252

Signed-off-by: Daniel Herman <dherman@factset.com>