fix(ui): add a tooltip for commonly truncated fields in the events pane
It's not uncommon for the involved object and/or message to be longer
than can be displayed in the events pane, and finding out what the full
string is has often involved going into the inspector to look at the contents
of the element.

Adding a tooltip on these will help mitigate that a bit.

Signed-off-by: Daniel Herman <dherman@factset.com>