fix: send periodic keepalive packets on eventstream connections
When operating behind load balancers / ingress controllers that have
an idle timeout configured, it's not uncommon to get disconnected and have
an error shown in the UI if you're looking at a relatively inactive workflow
or workflow list.

In the SSE spec, :\n is a sequence that you can send to the client which
should be ignored by the client, so we can use that to periodically send
something in the response without affecting the code in the UI at all.

Fixes #5006
Fixes #5124

Signed-off-by: Daniel Herman <dherman@factset.com>