Fix Leader Lease
When running multiple controllers within the same namespace if they share the lease lock name one of the will fail to start due to the metrics server not coming up and so the health check fails. Having unique names based off the LEADER_ELECTION_IDENTITY env variable should fix this and allow leaders election to work with multiple controllers in the same namespace.

Signed-off-by: Zach Aller <zachaller@hotmail.com>