feat(controller): Only send one event for pod nodes
When the workflow controller is configured to send events with the
`involvedObject` set to the pod for pod workflow nodes, it could either
only send the event with the pod as the involved object, or it could
send two events: one with the pod and the other as the workflow as the
involved object.

The design decision was made to only send a single event in order to
reduce unnecessary event noise. Users that opt-in to this feature must
be aware that pod nodes will no longer emit events with the Workflow as
the involved object.

Signed-off-by: Joseph McGovern <joseph.mcgovern@workiva.com>