Making metadata header lowercase to match selector
The name of the header in the metadata object inside the selector must be lowercase to match properly.  See the "Metadata" section below.  Trying with this example it was not matching, and thus not running the workflow.  Once I made the header name lowercase in the selector the selector matched and the workflow ran as expected.