fix: improve error message for ListArchivedWorkflows (#7345)
* fix: improve error message for ListArchivedWorkflows

Error message that is returned when user tries to list archived workflows and doesn't specify a namespace is very generic and hard to understand. This PR tries to make it a bit more clear as to what might be the cause of the issue, to help guide the user.

I thought that it might be a good idea to check if `namespace` is an empty string, and if it is, return a custom error for this case, telling user that `namespace` can't be empty. But I assume that this would not be correct for cases when Argo-Workflows runs with ClusterRole permissions?

Signed-off-by: Dominik Deren <dominik.deren@live.com>

* Fixing test & adding string formatting.

Signed-off-by: Dominik Deren <dominik.deren@live.com>