fix(ui): unauthorized login onboarding screen doesn't clarify required auth operations
My teammate @bulbul-ismail and I faced with an UNAUTH error while bring up our new argo-workflow-server pod in the internal network and after that tested the same template in our systems. Resolved the issue quickly and after that, we realized it would be a better instruction to print "provide the $ARGO_SERVER_ID via kubectl exec" in the login fail screen.

Signed-off-by: Doğukan Tuna <dogukan.tuna@trendyol.com>