fix: Fix inlined templates in templates
When a template is inline inside a WorkflowTemplate or
ClusterWorkflowTemplate we must still create a stored version of
it. This fix changes the 'name' given to those from
`/scope/templateName/` (where after the last / should be a name but
there isn't) to /scope/templateName/inline/name' where name can be
step name or task name.

I have added inline as it's possible to name an inline step the same
as a template in the same WorkflowTemplate.

Signed-off-by: Alan Clucas <alan@clucas.org>