Adjust gc_controller log to be more accurate
This log can not reflect the workflow which is to be deleted correctly. The call of Delete function returns nil does not mean the workflow has been successfully. We have seen it in production environment sometimes which is confused us. For example, if finalizers do not clean up, the workflow should  not be deleted.

Signed-off-by: peterandluc <peterandluc@gmail.com>