feat: mainEntrypoint variable
Issue #6772 asks for the template name, but we cannot always know what
this is (simplest unknowable is kubectl apply of a workflow with a
generateName). This is a split from PR #10745 to just provide the
variable mainEndpoint, which is set to the name of the invoked
entrypoint, which is controllable, knowable, and could be used in the
same way, and seems like a useful thing to have anyway.

Signed-off-by: Alan Clucas <alan@clucas.org>