fix: Allow hooks to be specified in workflowDefaults
Any hook in workflowDefaults will cause `Error: merging object in json
but data type is not struct, instead is: map`.

This is down to how StrategicMergePatch handles maps of objects and
limitations in merging them (a starting point for understading this is
https://github.com/golang/go/issues/33487).

Instead, we copy the map of hooks, patch it out of the patch, perform
the merge as before and then manually apply it.

fixes: #11095

Signed-off-by: Alan Clucas <alan@clucas.org>