fix: prevent memoization accessing wrong config-maps
Ensure that config-maps used for memoization aren't actually important
for other uses. Config-maps created for memoization would have the
label `workflows.argoproj.io/configmap-type: Cache`, but this was not
validated.

This commit changes that, and both load and saving to an existing
config-map will require that label to be present, or else the workflow
will error with a hopefully helpful message indicating the problem.

All existing config-maps which were created by memoization will have
this label already. The only concern is where a config-map existed
prior to the first save by memoization. Before this change the
config-map could still be used, and now workflows will error instead.

This improves security and prevents writing to the
`workflow-controller-configmap` by memoization.

I considered alternative paths for excluding
workflow-controller-configmap, such as a hardcoded list or a
configurable list, but this seems like the best longer term solution,
and requires zero configuration.

fixes: #5538

Signed-off-by: Alan Clucas <alan@clucas.org>