docs: Link memoization and work-avoidance
Despite #10769 and #10426 both having examples of memoization not
working with the examples having no output, no-one has picked up on
this.

To address this improve the documentation for memoization and
work-avoidance, linking the two ideas and pointing people who want to
skip steps towards work-avoidance unless they are really doing what
memoization was designed to do.

Issue #10426 is problematic in that some steps get memoized when
perhaps they should't, so this commit shouldn't close it.

Fixes #10769

Signed-off-by: Alan Clucas <alan@clucas.org>