docs: change quick-start guide to use quick-start install
The quick start guide used the regular install.yaml for installation
which does not configure the default user to be a valid service
account to execute a workflow. This causes any example
executions to fail. However the quick-start-minimal.yaml
installation does configure this and is clearly suited for this
doc so switch to that. fixes #12391

Signed-off-by: Doug Goldstein <cardoe@cardoe.com>