fix(resources): improve ressource accounting. Fixes #12468
The `.Values()` function rounds cpu ressources up to an integer. As a result, containers with small cpu requests (like the sidecar that requests 10m cores) are counted as requiring one full core, making the resources duration computation quite inaccurate.
This changes uses `.MilliValues()` instead. The `MilliValues()` documentation mentions that this function may overflow, but the return value is stored in an int64, so the overflow will happen if a container requests more than 9*10^(15) of a ressource. I assume this should be fine for a while.

Signed-off-by: Aloÿs Augustin <aloys.augustin@aqemia.com>