fix: don't fail workflow if PDB creation fails
It is likely that a busy kubernetes API will fail to create a PDB
occasionally. This results in the whole workflow being marked as
failed, when this is most likely something that will succeed on retry.

PDB creation only happens to workflows in state Unknown, so delay
marking the workflow is running until after we know if PDB creation
has succeeded.

If PDB creation fails then requeue the workflow and log the failure.

Signed-off-by: Alan Clucas <alan@clucas.org>