feat: OpenTelemetry metrics
Led by #12589.

As discussed in #12589 OpenTelemetry is the future of
observability. This PR changes the underlying codebase to collect
metrics using the otel libraries, whilst attempting to retain
compatibility with prometheus scraping where this makes sense.

It helps lay the groundwork for adding workflow tracing using otel.

This PR amends and extends the built in metrics to be more useful and
correctly named.

This PR does not attempt do do anything else with otel, nor does it
attempt to change custom metrics.

Mostly removed prometheus libraries, and replaced with otel libraries.

Allows use of both prometheus `/metrics` scraping and opentelemetry
protocol transmission.

Extends the workqueue metrics to include all of them exposed by
[client-go/workqueue](https://github.com/kubernetes/client-go/blob/babfe96631905312244f34500d2fd1a8a10b186c/util/workqueue/metrics.go#L173)

Removed `argo_workflows_workflows_processed_count` as a
non-functioning metric.

`originName` enhancements proposed in #12589 were not added, nor
`argo_workflows_workflowtemplate_used_total`

Note to reviewers: this is part of a stack of reviews for metrics
changes. Please don't merge until the rest of the stack is also ready.

Signed-off-by: Alan Clucas <alan@clucas.org>