feat: enable configuration of temporality in opentelemetry metrics
Implemented because of https://github.com/argoproj/argo-workflows/issues/12589#issuecomment-2174216603

Some backends (e.g. prometheus) only support cumulative temporality
while others (e.g. dynatrace) only support delta metrics.

It is possible to configure an opentelemetry collector between argo
workflow controller and the backend to convert cumulative metrics into
delta metrics but it has some limits (e.g. when controller is
restarted or when it is scaled up).

This commit enables the choice of temporality for OpenTelemetry

Temporality is always cumulative by definition in Prometheus

Note to reviewers: this is part of a stack of reviews for metrics
changes. Please don't merge until the rest of the stack is also ready.

Signed-off-by: Alan Clucas <alan@clucas.org>