feat: metrics configuration
Allows removal of metrics, removal of attributes (labels) and
reconfiguration of histogram buckets for all metrics using `modifiers`
in the workflow-controller-configmap.

Implemented using opentelemetry views, these modifiers work for
Prometheus metrics also

Note to reviewers: this is part of a stack of reviews for metrics
changes. Please don't merge until the rest of the stack is also ready.

Signed-off-by: Alan Clucas <alan@clucas.org>