fix(ui): hide `Workflow gone` message when workflow is archived
Fixes #13305

In workflows version 3.5 the workflow and archived views were merged.

When a workflow is deleted from the cluster the `Workflow gone`
message is shown in the UI to a user who is viewing the workflow. It
is still possible to interact with the workflow if it is archived as
though it was in cluster. This message is confusing to users.

With archiving on, delete a workflow which is archived. See that the
message is not shown.
With archiving on, allow the controller to delete a workflow which is
archived. See that the message is not shown
With archiving off, delete a workflow. See that the message is shown.

Signed-off-by: Alan Clucas <alan@clucas.org>