feat: new leader metric
Adds a leader metric which is 1 when this workflow controller is
actively running workflows.

With Prometheus output the non-leader doesn't emit metrics. This is
preserved in the new implementation, but OpenTelemetry does emit
metrics.

Using this metric you can supress or otherwise manipulate the metrics
from a non-leader controller if you would like to.

Note to reviewers: this is part of a stack of reviews for metrics
changes. Please don't merge until the rest of the stack is also ready.

Signed-off-by: Alan Clucas <alan@clucas.org>