fix: CronWorkflow schedules with timezone
Using `schedules` with a timezone specified does not work.
The cron schedule given to the cron backend includes the CRON_TZ=zone
string and should not.

The actual fix is one line to correctly honor the timezone flag on
`getSchedules` so that it doesn't double append when that flag is set,
nor single append when it is not. GetSchedules() would return an
invalid cron string resulting in the error:

```
could not schedule CronWorkflow" cronWorkflow=argo/go-forward error="expected exactly 5 fields, found 6: [CRON_TZ=America/Los_Angeles 30 2 * * *]
```

Added some more unit tests to check this behaves as expected, and a
smoke e2e test.

Signed-off-by: Alan Clucas <alan@clucas.org>