feat: UI can allow relative URL links (#13479)
Added the ability for users to supply relative links within the UI for
the links feature so that buttons and links do not have to have the
absolute URL baked into each link. This will allow for easily adding
buttons for different filters on the workflow list page for example.

Signed-off-by: Doug Goldstein <cardoe@cardoe.com>