feat(ui): Set namespace when submitting ClusterWorkflowTemplate
This adds support for specifying the namespace when submitting a
workflow from a ClusterWorkflowTemplate using an input field on the
"Submit Workflow".

Initially, I used the `<InputFilter>` component for this so it could
take advantage of autocomplete, but then I realized that's not
appropriate here because this isn't a filter and we don't want the "x"
button to clear the field.

Signed-off-by: Mason Malone <mmalone@adobe.com>